--- src/app.controller.ts
@@ -1,11 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { AppService } from './app.service';
-
-@Controller()
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-  @Get('/')
-  hello() {
-    return 'hello world';
-  }
-}
