+++ src/app.module.ts
@@ -2,20 +2,16 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { LightningModule } from './lightning/lightning.module';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
 import { UserModule } from './user/user.module';
 import { StudyModule } from './study/study.module';
 import { RedisModule } from './redis/redis.module';
 import { S3Module } from './s3/s3.module';
 import { ProfileModule } from './profile/profile.module';
-
 import { ChatsModule } from './chats/chats.module';
 import { EventsModule } from './events/events.module';
-
 import { StudyRecruitModule } from './study_recruit/study_recruit.module';
-import typeORMConfig from 'typeorm.config';
+import typeORMConfig from 'src/config/typeorm.config';
 import * as config from 'config';
 
 @Module({
@@ -36,7 +32,5 @@ import * as config from 'config';
     EventsModule,
     StudyRecruitModule,
   ],
-  controllers: [AppController],
-  providers: [AppService],
 })
 export class AppModule {}
