+++ src/auth/auth.module.ts
@@ -2,16 +2,16 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './services/auth.service';
 import { AuthController } from './controllers/auth.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthSocialLoginRepository } from './repositories/authSocialLogin.repository';
+import { AuthSocialLoginRepository } from './repositories/auth-social-login.repository';
 import { UserRepository } from 'src/user/repositories/user.repository';
-import { AuthPasswordLoginRepository } from './repositories/authPasswordLogin.repository';
+import { AuthPasswordLoginRepository } from './repositories/auth-password-login.repository';
 import { UserProfileRepository } from 'src/user/repositories/userProfile.repository';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import * as config from 'config';
-import { JwtStrategy } from './strategis/jwt.strategy';
+import { JwtStrategy } from '../config/strategis/jwt-access-token.strategy';
 import { RedisModule } from 'src/redis/redis.module';
-import JwtRefreshStrategy from './strategis/jwt-refresh_token.strategy';
+import JwtRefreshStrategy from '../config/strategis/jwt-refresh-token.strategy';
 import { UserImageRepository } from 'src/user/repositories/userImage.repository';
 import { AuthSocialService } from './services/auth-social.service';
 import { AuthSocialController } from './controllers/auth-social.controller';
