+++ src/lightning/controllers/lightning-recruitment.controller.ts
@@ -13,7 +13,7 @@ import { SuccessInterceptor } from 'src/common/interceptors/success.interceptor'
 import { HttpExceptionFilter } from 'src/common/exceptions/http-exception.filter';
 import { ApiOperation, ApiParam, ApiTags } from '@nestjs/swagger';
 import { LightningRecruitmentService } from '../services/lightning-recruitment.service';
-import { JwtAccessGuard } from 'src/auth/guards/jwt-access-token.guard';
+import { JwtAccessGuard } from 'src/config/guards/jwt-access-token.guard';
 
 @ApiTags('lightning-recruitment')
 @Controller('lightning-boards')
