+++ src/main.ts
@@ -1,19 +1,14 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { DocumentBuilder } from '@nestjs/swagger';
-import { OpenAPIObject, SwaggerModule } from '@nestjs/swagger';
 import { NestExpressApplication } from '@nestjs/platform-express';
-import * as expressBasicAuth from 'express-basic-auth';
 import * as config from 'config';
+import { setupSwagger } from './config/swagger';
+import { ValidationPipe } from '@nestjs/common';
 
 const serverConfig = config.get('server');
-const swaggerConfig = config.get('swagger');
-import { setupSwagger } from './util/swagger';
-import { ValidationPipe } from '@nestjs/common';
 
 async function bootstrap() {
   const app = await NestFactory.create<NestExpressApplication>(AppModule);
-  setupSwagger(app);
   app.useGlobalPipes(
     new ValidationPipe({
       transform: true,
@@ -23,31 +18,7 @@ async function bootstrap() {
     }),
   );
 
-  app.use(
-    ['/docs', '/docs-json'],
-    expressBasicAuth({
-      challenge: true,
-      users: {
-        [swaggerConfig.user]: swaggerConfig.password,
-      },
-    }),
-  );
-  const config = new DocumentBuilder()
-    .setTitle('5term-main-api')
-    .setDescription('모임서비스API')
-    .setVersion('1.0.0')
-    .addBearerAuth(
-      {
-        type: 'http',
-        scheme: 'bearer',
-        name: 'JWT',
-        in: 'header',
-      },
-      'access-token',
-    )
-    .build();
-  const document: OpenAPIObject = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('docs', app, document);
+  setupSwagger(app);
   app.enableCors({
     origin: true,
     credentials: true,
