+++ src/profile/profile.service.ts
@@ -4,8 +4,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { UserProfileRepository } from '../user/repositories/userProfile.repository';
 import { UserRepository } from 'src/user/repositories/user.repository';
 import { UserImageRepository } from 'src/user/repositories/userImage.repository';
-import { ChangePasswordDto } from 'src/auth/dtos/changePassword.dto';
-import { AuthPasswordLoginRepository } from 'src/auth/repositories/authPasswordLogin.repository';
+import { ChangePasswordDto } from 'src/auth/dtos/change-password.dto';
+import { AuthPasswordLoginRepository } from 'src/auth/repositories/auth-password-login.repository';
 import { ChangeEmailDto } from './dto/changeEmail.dto';
 import { ChangePhoneDto } from './dto/changePhone.dto';
 import { ChangeBioDto } from './dto/changeBio.dto';
