--- src/redis/redis.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RedisService } from './redis.service';
-
-describe('RedisService', () => {
-  let service: RedisService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [RedisService],
-    }).compile();
-
-    service = module.get<RedisService>(RedisService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
