+++ src/study/controller/study.controller.ts
@@ -12,8 +12,8 @@ import {
 } from '@nestjs/common';
 import { StudyService } from '../service/study.service';
 import { ApiOperation, ApiParam, ApiTags } from '@nestjs/swagger';
-import { GetUserId } from 'src/common/decorator/getUserId.decorator';
-import { JwtAccessGuard } from 'src/auth/guards/jwt-access-token.guard';
+import { GetUserId } from 'src/common/decorators/getUserId.decorator';
+import { JwtAccessGuard } from 'src/config/guards/jwt-access-token.guard';
 import { SuccessInterceptor } from 'src/common/interceptors/success.interceptor';
 import { HttpExceptionFilter } from 'src/common/exceptions/http-exception.filter';
 import { StudyCreateDto } from '../dto/study.create.dto';
