+++ src/user/repositories/user.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
 import { User } from '../entities/user.entity';
-import { IdDuplicationCheckDto } from 'src/auth/dtos/duplicationCheck.dto';
+import { IdDuplicationCheckDto } from 'src/auth/dtos/duplication-check.dto';
 
 @EntityRepository(User)
 export class UserRepository extends Repository<User> {
