--- src/util/swagger.ts
@@ -1,13 +1,0 @@
-import { INestApplication } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-
-export function setupSwagger(app: INestApplication): void {
-  const option = new DocumentBuilder()
-    .setTitle('NestJS Study API Docs')
-    .setDescription('NestJS Study API description')
-    .setVersion('1.0.0')
-    .build();
-
-  const document = SwaggerModule.createDocument(app, option);
-  SwaggerModule.setup('api-docs', app, document);
-}
